Data_Extraction/Fetch/reddit_reviews_agent/requirements.txt
praw==7.8.1
PyYAML==6.0.2
pytz==2024.1
tenacity==8.5.0
